+++ src/main/java/roomescape/domain/dao/JdbcReservationDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationDao;
+import roomescape.service.ReservationDao;
 
 @Repository
 public class JdbcReservationDao implements ReservationDao {
