+++ src/main/java/roomescape/domain/dao/JdbcReservationTimeDao.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimeDao;
+import roomescape.service.ReservationTimeDao;
 
 @Repository
 public class JdbcReservationTimeDao implements ReservationTimeDao {
