+++ src/main/java/roomescape/service/ReservationDao.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.service;
 
 import java.util.List;
 import roomescape.domain.Reservation;
