+++ src/main/java/roomescape/service/ReservationService.java
@@ -2,8 +2,6 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.ReservationDao;
-import roomescape.domain.ReservationTimeDao;
 import roomescape.domain.dto.ReservationRequest;
 import roomescape.domain.dto.ReservationResponse;
 import roomescape.domain.Reservation;
