+++ src/test/java/roomescape/service/FakeReservationTimeDao.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.ReservationTime;
-import roomescape.domain.ReservationTimeDao;
 
 public class FakeReservationTimeDao implements ReservationTimeDao {
